# storage/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Ticket Service Restful API",
        "description": "Documentation for the Application API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Event Ticket Service Restful API Server #1"
        }
    ],
    "paths": {
...
        "/v1/reservations/{reservation}": {
            "get": {
                "summary": "Get a reservation",
                "description": "Get a reservation",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "Reservation retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "Reservations"
                ],
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "required": true,
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "summary": "Update a reservation",
                "description": "Update a reservation",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "Reservation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "Reservations"
                ],
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "required": true,
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "summary": "Update a reservation",
                "description": "Update a reservation",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "Reservation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "Reservations"
                ],
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "required": true,
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete a reservation",
                "description": "Delete a reservation",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "Reservation deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "Reservations"
                ],
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "required": true,
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "schemas": {
...
            "Reservation": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "user_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "event_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "number_of_tickets": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": true,
                        "description": "int unsigned"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "required": [
                    "id",
                    "user_id",
                    "event_id",
                    "number_of_tickets"
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}